+++ app/index.js
@@ -1,93 +1,141 @@
-"use strict";
-var __awaiter = (this && this.__awaiter) || function (thisArg, _arguments, P, generator) {
-    function adopt(value) { return value instanceof P ? value : new P(function (resolve) { resolve(value); }); }
-    return new (P || (P = Promise))(function (resolve, reject) {
-        function fulfilled(value) { try { step(generator.next(value)); } catch (e) { reject(e); } }
-        function rejected(value) { try { step(generator["throw"](value)); } catch (e) { reject(e); } }
-        function step(result) { result.done ? resolve(result.value) : adopt(result.value).then(fulfilled, rejected); }
-        step((generator = generator.apply(thisArg, _arguments || [])).next());
-    });
-};
-var __generator = (this && this.__generator) || function (thisArg, body) {
-    var _ = { label: 0, sent: function() { if (t[0] & 1) throw t[1]; return t[1]; }, trys: [], ops: [] }, f, y, t, g = Object.create((typeof Iterator === "function" ? Iterator : Object).prototype);
-    return g.next = verb(0), g["throw"] = verb(1), g["return"] = verb(2), typeof Symbol === "function" && (g[Symbol.iterator] = function() { return this; }), g;
-    function verb(n) { return function (v) { return step([n, v]); }; }
-    function step(op) {
-        if (f) throw new TypeError("Generator is already executing.");
-        while (g && (g = 0, op[0] && (_ = 0)), _) try {
-            if (f = 1, y && (t = op[0] & 2 ? y["return"] : op[0] ? y["throw"] || ((t = y["return"]) && t.call(y), 0) : y.next) && !(t = t.call(y, op[1])).done) return t;
-            if (y = 0, t) op = [op[0] & 2, t.value];
-            switch (op[0]) {
-                case 0: case 1: t = op; break;
-                case 4: _.label++; return { value: op[1], done: false };
-                case 5: _.label++; y = op[1]; op = [0]; continue;
-                case 7: op = _.ops.pop(); _.trys.pop(); continue;
-                default:
-                    if (!(t = _.trys, t = t.length > 0 && t[t.length - 1]) && (op[0] === 6 || op[0] === 2)) { _ = 0; continue; }
-                    if (op[0] === 3 && (!t || (op[1] > t[0] && op[1] < t[3]))) { _.label = op[1]; break; }
-                    if (op[0] === 6 && _.label < t[1]) { _.label = t[1]; t = op; break; }
-                    if (t && _.label < t[2]) { _.label = t[2]; _.ops.push(op); break; }
-                    if (t[2]) _.ops.pop();
-                    _.trys.pop(); continue;
+import express from "express" ; 
+import   bcrypt from 'bcrypt' ; 
+import  jwt from "jsonwebtoken" 
+import * as mongoose from 'mongoose' 
+import * as zod from 'zod' ; 
+import { userModel } from "./schema/user.schema.js" ; 
+import { contentModel } from "./schema/content.schema.js";
+import connectDB from "./db.js"
+import { userMiddleware } from "./middleware/auth.middleware.js";
+import signUpZodSchema from "./zod_schema/signUp.zod.js";
+import dotenv from 'dotenv' ; 
+dotenv.config("") ; 
+const app = express() ; 
+app.use(express.json()) ; 
+
+connectDB() ; 
+
+app.get("/" , (req, res)=>{ 
+    res.send("hello world !")
+})
+
+app.post("/api/v1/signin", async (req , res)=>{
+     
+    try { 
+        const validateInputData = signUpZodSchema.parse(req.body) ; 
+        const { username , password } = validateInputData ; 
+        const existingUser = await userModel.findOne({
+            username 
+        }) ; 
+        
+        if(existingUser) { 
+            const isPasswordValid = bcrypt.compare(password , existingUser.password) ; 
+
+            if(!isPasswordValid) { 
+                res.json({
+                    message : "Incorrect Password Entered! Please try again ! "
+                }).status(501) ; 
             }
-            op = body.call(thisArg, _);
-        } catch (e) { op = [6, e]; y = 0; } finally { f = t = 0; }
-        if (op[0] & 5) throw op[1]; return { value: op[0] ? op[1] : void 0, done: true };
-    }
-};
-Object.defineProperty(exports, "__esModule", { value: true });
-var express = require("express");
-var user_schema_js_1 = require("./schema/user.schema.js");
-var db_1 = require("./db");
-var app = express();
-app.use(express.json());
-(0, db_1.default)();
-app.get("/", function (req, res) {
-    res.send("hello world !");
-});
-app.post("/api/v1/signin", function (req, res) {
-});
-app.post("/api/v1/signup", function (req, res) { return __awaiter(void 0, void 0, void 0, function () {
-    var username, password, e_1;
-    return __generator(this, function (_a) {
-        switch (_a.label) {
-            case 0:
-                username = req.body.username;
-                password = req.body.password;
-                _a.label = 1;
-            case 1:
-                _a.trys.push([1, 3, , 4]);
-                return [4 /*yield*/, user_schema_js_1.userModel.create({
-                        username: username,
-                        password: password,
-                    })];
-            case 2:
-                _a.sent();
-                res.json({
-                    message: "user ".concat(username, " created succssfully !"),
-                    status: 201,
-                }).status(201);
-                console.log("user created successfully ! ");
-                return [3 /*break*/, 4];
-            case 3:
-                e_1 = _a.sent();
-                res.json({
-                    message: "Error occurred while creating user "
-                }).status(501);
-                console.log(e_1);
-                return [3 /*break*/, 4];
-            case 4: return [2 /*return*/];
+
+            const token = jwt.sign({
+                id : existingUser._id, 
+            }, `${process.env.JWT_TOKEN_SECRET}`) ; 
+            
+            res.json({
+                message : "Sign in successful"  , 
+                token : token , 
+            })
+        }else { 
+            res.json({
+                message : "Incorrect credentials" ,
+            }).status(403)
         }
-    });
-}); });
-app.get("/api/v1/content", function (req, res) {
-});
-app.delete("/api/v1/content", function (req, res) {
-});
-app.post("/api/v1/brain/share", function (req, res) {
-});
-app.get("/api/v1/brain/:shareLink", function (req, res) {
-});
-app.listen(3000, function () {
-    console.log("server started");
-});
+    }catch(e) { 
+        if (e instanceof zod.ZodError)  { 
+            res.json({
+                message : e.message ,
+            })
+        }
+        console.log("Error occurred while finding the user") ; 
+        res.json({
+            message : "Error occurred while finding the user" 
+        }).status(501)  ; 
+    }
+
+} )
+
+app.post("/api/v1/signup", async  (req , res )=>{
+    try {
+        const validateInputData = signUpZodSchema.parse(req.body); 
+        const { username , password } = validateInputData ; 
+        
+        const hashedPassword = await bcrypt.hash(password , 10) ; 
+        await userModel.create({
+            username : username , 
+            password : hashedPassword  , 
+        })
+
+        res.json({
+            message : `user ${username} created succssfully !` ,
+            status : 201 ,  
+        }).status(201) 
+        console.log("user created successfully ! ") ; 
+    } 
+    catch(e) { 
+        if (e instanceof zod.ZodError)  { 
+            res.json({
+                message : e.message.message ,
+            })
+        }
+        res.json({
+            message : `Error occurred while creating user `
+        }).status(501) ;  
+        console.log(e)
+    }
+}) 
+
+app.post("/api/v1/content" , userMiddleware, async  (req , res)=>{
+    const { link , title , type, tag , userId } = req.body ;  
+
+    try {
+        await contentModel.create({
+            link : link , 
+            title : title , 
+            type : type , 
+            tag : tag , 
+            userId : userId 
+        }) 
+        res.json({
+            message : "Content created succesfully ! "
+        }).status(201)   
+
+    }
+    catch(e) { 
+        console.log("Error occurred while creating content ") 
+        res.json({
+            message: "error while creating content"
+        }).status(501) ; 
+    }
+})
+
+
+app.get("/api/v1/content" , (req , res )=>{
+
+})
+
+app.delete("/api/v1/content" , (req , res )=>{
+
+})
+
+app.post("/api/v1/brain/share" , (req ,res )=>{
+
+})
+
+app.get("/api/v1/brain/:shareLink" , (req , res)=>{
+
+})
+
+
+app.listen(3000, ()=>{
+    console.log("server started")
+})
