+++ app/schema/content.schema.js
