+++ app/schema/user.schema.js
@@ -1,16 +1,15 @@
-"use strict";
-Object.defineProperty(exports, "__esModule", { value: true });
-exports.userModel = void 0;
-var mongoose_1 = require("mongoose");
-var userSchema = new mongoose_1.default.Schema({
-    username: {
-        required: true,
-        type: String,
-        unique: true,
-    },
-    password: {
-        required: true,
-        type: String,
+import mongoose from 'mongoose' ;
+
+const userSchema = new mongoose.Schema({
+    username : {
+        required : true,  
+        type : String  , 
+        unique : true , 
+    }, 
+    password : { 
+        required : true , 
+        type : String , 
     }
-});
-exports.userModel = mongoose_1.default.model("user", userSchema);
+})
+
+export const userModel = mongoose.model("user", userSchema) ;
